requirements.txt
couchbase==2.2.1
coverage==4.4.1
eventlet==0.21.0
flake8==3.3.0
Flask==0.12.2
isort==4.2.5
matplotlib==2.0.2
nose==1.3.7
numpy==1.12.1
pandas==0.20.2
pep8-naming==0.4.1
pillow==4.1.1
pydocstyle==2.0.0
